+++ src/test/java/nextstep/subway/auth/session/SessionAuthenticationInterceptorTest.java
@@ -6,7 +6,7 @@ import nextstep.subway.auth.ui.common.AuthenticationConverter;
 import nextstep.subway.auth.ui.session.SessionAuthenticationConverter;
 import nextstep.subway.auth.ui.session.SessionAuthenticationInterceptor;
 import nextstep.subway.member.application.CustomUserDetailsService;
-import nextstep.subway.member.application.UserDetailService;
+import nextstep.subway.auth.application.UserDetailService;
 import nextstep.subway.member.domain.LoginMember;
 
 import org.junit.jupiter.api.BeforeEach;
