+++ src/test/java/nextstep/subway/auth/token/TokenAuthenticationInterceptorTest.java
@@ -9,7 +9,7 @@ import nextstep.subway.auth.ui.common.AuthenticationConverter;
 import nextstep.subway.auth.ui.token.TokenAuthenticationConverter;
 import nextstep.subway.auth.ui.token.TokenAuthenticationInterceptor;
 import nextstep.subway.member.application.CustomUserDetailsService;
-import nextstep.subway.member.application.UserDetailService;
+import nextstep.subway.auth.application.UserDetailService;
 import nextstep.subway.member.domain.LoginMember;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
